refactor(goesr): update import for GOESPlatformMetadata and move it to netcdf_info.py
