+++ src/goesdl/goesr/netcdf_info.py
